backend_implementation/requirements.txt
# Contract Generation Backend Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
flask>=3.0.0
python-docx>=1.1.0

# If you're using additional frameworks (uncomment as needed):
# flask-cors>=4.0.0
# flask-limiter>=3.5.0

# Database (if tracking contracts):
# psycopg2-binary>=2.9.0  # PostgreSQL
# sqlalchemy>=2.0.0

# LLM API (uncomment based on your provider):
# openai>=1.0.0
# anthropic>=0.7.0

# Testing
# pytest>=7.4.0
# pytest-flask>=1.3.0
